Extract compile and run options into dedicated files
